+++ src/translation/fr-FR/components/DataDisplay/Error/index.ts
@@ -1,0 +1,5 @@
+import {IKitErrorLocale} from '@translation/types/components/DataDisplay/Error';
+
+export const KitErrorLocale: IKitErrorLocale = {
+    error404: 'Type d’erreur : 404'
+};